Trust the root CA of the *.typo3.org certificate on Debian Squeeze
The root CA certificate used there is not trusted on Debian Squeeze.
As some of our servers are still running with this ancient OS, tools
like `curl` wouldn't be able to contact other *.typo3.org servers,
as the certificate validation fails.

Therefore, manually add the root CA certificate to the trust store.