Define the production state explilictly through a flag instead of automatism
Previously, we detected the fact that we are running in production by
searching the `datacenter` data bag for our (host name OR our virtualization
host's name).

But: This lead to a chicken-egg problem during bootstrapping.
(in other words: it worked only for existing nodes)

> We are not running in a data center
> Then we do not apply the chef-client configuration
> Which includes our magic custom ohai plugin to detect the host system [1].
> When we cannot find the data center, in which our host system is running
> Then we are not including the `_production` recipe
> Which would allow us to detect our virtualization host.

This change goes along with 6d2284373d in the main chef-repo, which sets
this flag in the `production` environment.

[1] https://github.com/TYPO3-cookbooks/t3-openvz/blob/8420c0cee25d376282ffd783d8e2b3edd9d55814/templates/default/ohai/openvz-hostdetection.rb