[TASK] Give the vagrant user sudo rights
kitchen-vagrant seems to use the `root` user in Debian boxes,
but the `vagrant` user in Ubuntu boxes (thus needs sudo).
kitchen-docker uses the `kitchen` user, for which we already have
a sudo setup.

While we are not using Ubuntu for production, it becomes handy to
not pro-actively destroy sudo access for the user running kitchen,
e.g., if one wants to cross-check a community cookbook that seems
to not work well on Debian and figure out if it's the same for
Ubuntu.

Therefore, set up sudo correctly for both users, `kitchen` and
`vagrant`.